requirements.txt
argon==0.2.0
argon2==0.1.10
argon2-cffi==18.1.0
atomicwrites==1.1.5
attrs==18.1.0
backcall==0.1.0
bcrypt==3.1.4
cairocffi==0.9.0
CairoSVG==2.2.1
certifi==2018.4.16
cffi==1.11.5
coverage==4.5.1
cssselect2==0.2.1
decorator==4.3.0
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.0.8
django-bootstrap3==10.0.1
django-bootstrap4==0.0.6
django-filter==2.0.0
django-jet==1.0.7
django-widget-tweaks==1.4.2
djangorestframework==3.8.2
environ==1.0
html5lib==1.0.1
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.1
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.4.0
Markdown==2.6.11
misaka==2.1.0
more-itertools==4.2.0
nbformat==4.4.0
nbval==0.9.1
parso==0.3.1
pdfrw==0.4
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
pluggy==0.6.0
prompt-toolkit==1.0.15
ptyprocess==0.6.0
py==1.5.4
pycparser==2.18
Pygments==2.2.0
Pyphen==0.9.5
pytest==3.6.3
python-dateutil==2.7.3
python-decouple==3.1
pytz==2018.4
pyzmq==17.1.0
simplegeneric==0.8.1
six==1.11.0
tinycss2==0.6.1
tornado==5.1
traitlets==4.3.2
wcwidth==0.1.7
WeasyPrint==0.42.3
webencodings==0.5.1
